implement some refactor
