+++ modules/json_storage.rb
