+++ storageFiles/genre.json
@@ -1,4 +1,3 @@
 [
-  "pop",
-  "hiphop"
+  "pop"
 ]
